Add files via upload
This adds simulation features including a main scene that contains the model used in the scene with a body of water. Also, it adds movement and steering to the simulation including a rotation and movement system for forward and backward. Plus, there is a shader that was programmed through a tutorial giving the body of water noise textures to animate it for now. It also contains a camera to view the player and prototype controls for the camera including left and right movement with up and down movement.