Fixed kayak movement
This code is designed to make the kayak model move forward and backward in its current rotation. This works correctly now. The reason for the x-axis movement is that the model was imported from Blender, and in Blender, the orientation of the model is completely different from what Godot should be so x-axis could be forward and is in this case